+++ mobile/src/components/notes/index.ts
@@ -2,3 +2,5 @@ export { default as NoteItem } from './NoteItem';
 export { default as NotesMenu } from './NotesMenu';
 export { default as SelectionMenu } from './SelectionMenu';
 export { default as SelectionTitleBar } from './SelectionTitleBar';
+
+export { NOTES_PLACEHOLDERS } from '@/constants/data';
