+++ mobile/src/components/onboarding/index.ts
@@ -4,4 +4,4 @@ export { default as OnboardingSlide } from './OnboardingSlide';
 export { default as ThemeSelector } from './ThemeSelector';
 export { default as WelcomeIllustration } from './WelcomeIllustration';
 
-export { onboardingSteps, type OnboardingStepProps } from '@/constants/onboardingSteps';
+export { onboardingSteps, type OnboardingStepProps } from '@/constants/data';
