+++ mobile/src/constants/constants.ts
@@ -3,3 +3,14 @@ export const DATABASE_NAME = 'checklists';
 export const MAX_INPUT_LENGTH = 50;
 
 export const PRIVACY_POLICY_URL = 'https://checked.kriteshtimsina.com.np';
+
+export const NOTES_PLACEHOLDERS = [
+  'Untitled brilliance',
+  'Big brain moment',
+  'Random thoughts',
+  '// TODO: write something useful',
+  'Future you will thank you',
+  'Another genius idea',
+  'Capture the idea',
+  'Why did I open this…',
+];
