+++ mobile/src/constants/data.ts
@@ -1,78 +1,62 @@
-export const data = [
-  {
-    title: 'App deployment checklist',
-    createdAt: 'Feb 2',
-    checklist: [
-      {
-        title: 'Change app icon',
-        completed: false,
-        id: 1,
-      },
-      {
-        title: 'Change splash screen',
-        completed: false,
-        id: 2,
-      },
-      {
-        title: 'Update build version in app/build/build.gradle',
-        completed: false,
-        id: 3,
-      },
-      {
-        title: 'Change version number in package.json',
-        completed: false,
-        id: 4,
-      },
-      {
-        title: 'Deploy',
-        completed: false,
-        id: 5,
-      },
-    ],
-  },
+export const NOTES_PLACEHOLDERS = [
+  'Untitled brilliance',
+  'Big brain moment',
+  'Random thoughts',
+  '// TODO: write something useful',
+  'Future you will thank you',
+  'Another genius idea',
+  'Capture the idea',
+  'Why did I open this…',
 ];
 
-export const colors = [
+export type OnboardingStepProps = {
+  id: number;
+  emoji: string;
+  tag: string;
+  title: string;
+  titleAccent: string;
+  subtitle: string;
+  color: string;
+  accent: string;
+  bg: string;
+  cta: string;
+};
+
+export const onboardingSteps: OnboardingStepProps[] = [
+  {
+    id: 0,
+    emoji: '👋',
+    tag: 'Welcome',
+    title: 'Your brain,',
+    titleAccent: 'organized.',
+    subtitle: 'A fun, colorful space to capture every idea, task, and habit — all in one place.',
+    color: '#FF6B6B',
+    accent: '#FFE66D',
+    bg: '#FFF5F5',
+    cta: "What's next",
+  },
   {
     id: 1,
-    primary: '#e0c59e',
-    secondary: '#EAF8FF',
-    selected: 1,
+    emoji: '📝',
+    tag: 'Notes',
+    title: 'Capture every',
+    titleAccent: 'big idea.',
+    subtitle: 'Jot down thoughts, doodles, lists — your notes are always colorful, always yours.',
+    color: '#C77DFF',
+    accent: '#FF9F1C',
+    bg: '#FAF5FF',
+    cta: 'Sounds good',
   },
   {
     id: 2,
-    primary: '#22c55e',
-    secondary: '#EAF8FF',
-    selected: 0,
-  },
-  {
-    id: 3,
-    primary: '#8b5cf6',
-    secondary: '#EAF8FF',
-    selected: 0,
-  },
-  {
-    id: 4,
-    primary: '#f43f5e',
-    secondary: '#EAF8FF',
-    selected: 0,
-  },
-  {
-    id: 5,
-    primary: '#06b6d4',
-    secondary: '#EAF8FF',
-    selected: 0,
-  },
-];
-export const tabs = [
-  {
-    id: 1,
-    title: 'Tasks',
-    label: 'index',
-  },
-  {
-    id: 2,
-    title: 'Notes',
-    label: 'notes',
+    emoji: '🎨',
+    tag: 'Make it yours',
+    title: 'Choose your',
+    titleAccent: 'style.',
+    subtitle: 'Pick a theme that feels like you. You can always change it later in settings.',
+    color: '#B8864E',
+    accent: '#7A5230',
+    bg: '#FDF8F2',
+    cta: "Let's go →",
   },
 ];
