--- mobile/src/constants/notes.ts
@@ -1,10 +1,0 @@
-export const notesPlaceholders = [
-  'Untitled brilliance',
-  'Big brain moment',
-  'Random thoughts',
-  '// TODO: write something useful',
-  'Future you will thank you',
-  'Another genius idea',
-  'Capture the idea',
-  'Why did I open this…',
-];
